refactor: move refactoring logic to processor
